remove useApiWithProfile
just call mutations directly from components
