+++ src/pages/CreateCirclePage/CreateCircleForm.tsx
@@ -12,7 +12,7 @@ import { z } from 'zod';
 
 import { FormAutocomplete, FormInputField } from 'components';
 import { QUERY_KEY_MAIN_HEADER } from 'components/MainLayout/getMainHeaderData';
-import { useApiWithProfile } from 'hooks';
+import { useToast } from 'hooks';
 import { QUERY_KEY_MY_ORGS } from 'pages/CirclesPage/getOrgData';
 import { paths } from 'routes/paths';
 import {
@@ -29,6 +29,7 @@ import {
 import { SingleColumnLayout } from 'ui/layouts';
 
 import { CreateSampleCircle } from './CreateSampleCircle';
+import { createCircleMutation } from './mutations';
 import { CreateCircleQueryData } from './queries';
 
 export const NEW_CIRCLE_CREATED_PARAMS = '?new-circle';
@@ -61,6 +62,7 @@ export const CreateCircleForm = ({
 }) => {
   const navigate = useNavigate();
   const [params] = useSearchParams();
+  const { showError } = useToast();
 
   const queryClient = useQueryClient();
 
@@ -72,7 +74,6 @@ export const CreateCircleForm = ({
     avatarRaw: null,
   });
 
-  const { createCircle } = useApiWithProfile();
   const myUsers = source.myUsers;
   const organizations = useMemo(
     () =>
@@ -103,12 +104,13 @@ export const CreateCircleForm = ({
       const image_data_base64 = logoData.avatarRaw
         ? await fileToBase64(logoData.avatarRaw)
         : undefined;
-      const newCircle = await createCircle({
+      const newCircle = await createCircleMutation({
         ...data,
         image_data_base64,
       });
       circleCreated(newCircle.id);
     } catch (e) {
+      showError(e);
       console.warn(e);
     }
     reset(data);
